debt: playground: extract connection.js (mode switching + overlays)
The biggest single section lifted out — 395 lines of Data Source,
connection-status, overlay-rendering, LiveSource lifecycle, and
live/sim mode switching. Owns liveSource, simSource,
connectionStatus, lastDataTime, stalenessTimer — all module-local
now. main.js wires in a setRunning callback via initConnection so
switchToLive can still pause the simulation.

Also exports handleOverrideResponse from relay-board.js (previously
internal) so the new ensureLiveSource can attach it on LiveSource
creation.

balance-card.js: new resetBalanceState() centralizes the
balanceHistory/points/events/lastMode reset previously inlined
(orphaned after balance state moved).

main.js: 1091 → 699 lines. npm test:unit 782/782 pass. knip clean
after dropping the unused getSimSource export + de-exporting
helpers used only within connection.js.
