+++ playground/js/main.js
@@ -44,405 +44,12 @@ import {
   setSchematicHandle, getLastFrame,
   setLiveFrameSeen, resetYesterdayTracking,
 } from './main/display-update.js';
+import {
+  initConnection, initModeToggle, updateSidebarSubtitle, getLiveSource,
+} from './main/connection.js';
 // Expose for e2e testing
 window.__triggerVersionCheck = triggerVersionCheck;
 
-// ── Data Source ──
-// Detect deployment context: GitHub Pages = simulation only, deployed app = live capable
-// `window.__simulateGitHubPagesDeploy` is a test-only hatch (set via
-// Playwright's addInitScript) — it lets the e2e suite exercise the
-// pages-mode auto-bootstrap path without needing a real github.io URL.
-const isGitHubPages = window.__simulateGitHubPagesDeploy === true
-  || location.hostname.endsWith('.github.io');
-const isLocalhost = location.hostname === 'localhost' || location.hostname === '127.0.0.1';
-const isLiveCapable = !isGitHubPages;
-const urlModePreference = new URLSearchParams(location.search).get('mode'); // 'sim' | 'live' | null
-store.set('isLiveCapable', isLiveCapable);
-
-// activeSource is eliminated — use store.get('phase') instead
-let liveSource = null;
-let simSource = new SimulationSource();
-
-// Connection status tracking
-let connectionStatus = 'disconnected';
-let lastDataTime = 0;
-let stalenessTimer = null;
-
-function updateConnectionUI(status) {
-  connectionStatus = status;
-  refreshConnectionIndicator();
-  updateConnectionOverlays();
-  updateDevicePushState();
-}
-
-function refreshConnectionIndicator() {
-  const dot = document.getElementById('connection-dot');
-  const label = document.getElementById('connection-label');
-  if (!dot || !label) return;
-
-  if (connectionStatus === 'reconnecting') {
-    dot.className = 'connection-dot reconnecting';
-    label.textContent = 'Reconnecting\u2026';
-    return;
-  }
-
-  var displayState = getConnectionDisplayState();
-  switch (displayState) {
-    case 'active':
-      dot.className = 'connection-dot connected';
-      label.textContent = 'Live';
-      break;
-    case 'connecting':
-      dot.className = 'connection-dot reconnecting';
-      label.textContent = 'Connecting\u2026';
-      break;
-    case 'device_offline':
-      dot.className = 'connection-dot device-offline';
-      label.textContent = 'Controller offline';
-      break;
-    case 'stale':
-      dot.className = 'connection-dot reconnecting';
-      label.textContent = 'Stale';
-      break;
-    default:
-      dot.className = 'connection-dot disconnected';
-      label.textContent = 'Offline';
-  }
-}
-
-function checkStaleness() {
-  if (store.get('phase') !== 'live') return;
-  const banner = document.getElementById('staleness-banner');
-  if (banner) {
-    const elapsed = Date.now() - lastDataTime;
-    banner.classList.toggle('visible', elapsed > 60000 && lastDataTime > 0);
-  }
-  refreshConnectionIndicator();
-  updateConnectionOverlays();
-  updateDevicePushState();
-  updateSidebarSubtitle();
-}
-
-var OVERLAY_MESSAGES = {
-  connecting: {
-    title: 'Reaching out to your sanctuary.',
-    subtitle: 'Connecting to the server...'
-  },
-  never_connected: {
-    title: 'Your sanctuary is sleeping.',
-    subtitle: 'Cannot reach the server.'
-  },
-  device_offline: {
-    title: 'Your sanctuary is sleeping.',
-    subtitle: 'The server is running, but the controller is unreachable.'
-  },
-  disconnected: {
-    title: 'Lost touch with your sanctuary.',
-    subtitle: 'Connection to the server was lost.'
-  },
-  stale: {
-    title: 'Your sanctuary has gone quiet.',
-    subtitle: 'No data received for over 60 seconds.'
-  }
-};
-
-function getConnectionDisplayState() {
-  if (store.get('phase') !== 'live') return 'active';
-  var hasData = liveSource && liveSource.hasReceivedData;
-  var mqttStatus = liveSource ? liveSource.mqttStatus : 'unknown';
-
-  if (connectionStatus === 'connected') {
-    // WS is open — check MQTT and device data status
-    if (mqttStatus === 'disconnected' || mqttStatus === 'reconnecting') {
-      return 'device_offline';
-    }
-    if (hasData) {
-      if (lastDataTime > 0 && (Date.now() - lastDataTime) > 60000) return 'stale';
-      return 'active';
-    }
-    // WS connected, waiting for MQTT status or data — show connecting overlay
-    // Brief grace period (2s) before declaring device offline
-    var connectedAt = liveSource ? liveSource._connectedAt : 0;
-    if (connectedAt > 0 && (Date.now() - connectedAt) > 2000) {
-      return 'device_offline';
-    }
-    return 'connecting';
-  }
-
-  // WS is not connected or reconnecting
-  var wsEverFailed = liveSource && liveSource._wsEverFailed;
-  if (!hasData) return wsEverFailed ? 'never_connected' : 'connecting';
-  return 'disconnected';
-}
-
-function updateConnectionOverlays() {
-  var state = getConnectionDisplayState();
-  var overlayIds = ['overlay-modes', 'overlay-gauge', 'overlay-components'];
-  var msg = OVERLAY_MESSAGES[state];
-  for (var i = 0; i < overlayIds.length; i++) {
-    var overlay = document.getElementById(overlayIds[i]);
-    if (!overlay) continue;
-    if (msg) {
-      overlay.classList.add('visible');
-      var titleEl = document.getElementById(overlayIds[i] + '-title');
-      var subtitleEl = document.getElementById(overlayIds[i] + '-subtitle');
-      if (titleEl) titleEl.textContent = msg.title;
-      if (subtitleEl) subtitleEl.textContent = msg.subtitle;
-    } else {
-      overlay.classList.remove('visible');
-    }
-  }
-}
-
-function updateDevicePushState() {
-  var btn = document.getElementById('dc-save');
-  var warning = document.getElementById('dc-connection-warning');
-  if (!btn || !warning) return;
-  var displayState = getConnectionDisplayState();
-  var canPush = store.get('phase') !== 'live' || displayState === 'active' || displayState === 'stale';
-  if (canPush) {
-    btn.classList.remove('disabled');
-    btn.disabled = false;
-    warning.style.display = 'none';
-  } else {
-    btn.classList.add('disabled');
-    btn.disabled = true;
-    warning.style.display = '';
-  }
-}
-
-function initModeToggle() {
-  const toggle = document.getElementById('mode-toggle');
-  const sw = document.getElementById('mode-toggle-switch');
-  const label = document.getElementById('mode-toggle-label');
-  if (!toggle || !isLiveCapable) {
-    store.set('phase', 'simulation');
-    return;
-  }
-
-  toggle.classList.add('visible');
-  // Respect URL mode preference, default to live on deployed app
-  if (urlModePreference === 'sim') {
-    switchToSimulation();
-  } else {
-    switchToLive();
-  }
-
-  sw.addEventListener('click', function () {
-    if (store.get('phase') === 'live') {
-      switchToSimulation();
-    } else {
-      switchToLive();
-    }
-  });
-}
-
-function persistModeInUrl(mode) {
-  var params = new URLSearchParams(location.search);
-  if (mode === 'live') {
-    params.delete('mode');
-  } else {
-    params.set('mode', mode);
-  }
-  var qs = params.toString();
-  var url = location.pathname + (qs ? '?' + qs : '') + location.hash;
-  history.replaceState(null, '', url);
-}
-
-function updateModeToggleUI(isLive) {
-  const sw = document.getElementById('mode-toggle-switch');
-  const label = document.getElementById('mode-toggle-label');
-  if (isLive) {
-    if (sw) sw.classList.add('active');
-    if (label) { label.textContent = 'Live'; label.classList.add('active'); }
-  } else {
-    if (sw) sw.classList.remove('active');
-    if (label) { label.textContent = 'Simulation'; label.classList.remove('active'); }
-  }
-}
-
-function ensureLiveSource() {
-  if (!liveSource) {
-    liveSource = new LiveSource();
-    liveSource.onCommandResponse(handleOverrideResponse);
-    liveSource.onUpdate(function (state, result) {
-      lastDataTime = Date.now();
-      if (store.get('phase') !== 'live') return;
-      setLiveFrameSeen(true);
-      // Defense-in-depth: each step is independent. A bug in updateDisplay
-      // must not break the manual override controls (or vice versa).
-      var steps = [
-        ['updateDisplay',           function () { updateDisplay(state, result); }],
-        ['refreshConnectionIndicator', refreshConnectionIndicator],
-        ['updateConnectionOverlays',   updateConnectionOverlays],
-        ['updateSidebarSubtitle',      updateSidebarSubtitle],
-        ['updateDevicePushState',      updateDevicePushState],
-        ['updateRelayBoard',        function () { updateRelayBoard(result); }],
-      ];
-      for (var i = 0; i < steps.length; i++) {
-        try { steps[i][1](); }
-        catch (e) { console.error('liveSource.onUpdate ' + steps[i][0] + ' failed:', e); }
-      }
-    });
-    liveSource.onConnectionChange(function (status) {
-      if (store.get('phase') === 'live') {
-        updateConnectionUI(status);
-        updateSidebarSubtitle();
-      }
-    });
-    // Re-attach the watchdog broadcast listener now that liveSource exists.
-    attachWatchdogWebSocket();
-    attachScriptStatusWebSocket(liveSource);
-  }
-  liveSource.start();
-}
-
-function startStalenessTimer() {
-  if (!stalenessTimer) {
-    stalenessTimer = setInterval(checkStaleness, 5000);
-  }
-}
-
-function stopStalenessTimer() {
-  if (stalenessTimer) { clearInterval(stalenessTimer); stalenessTimer = null; }
-}
-
-function switchToLive() {
-  store.set('phase', 'live');
-  updateModeToggleUI(true);
-  persistModeInUrl('live');
-  ensureLiveSource();
-  running = false;
-  fetchLiveHistory(graphRange);
-  fetchLiveEvents(null);
-  fetchBalanceHistory();
-  startStalenessTimer();
-  clearLiveDisplay();
-  updateConnectionUI('disconnected');
-  updateSidebarSubtitle();
-  updatePhaseAwareCopy();
-}
-
-function clearLiveDisplay() {
-  // Reset display to placeholder values — never show simulation defaults in live mode
-  setLiveFrameSeen(false);
-  document.getElementById('mode-card-title').textContent = '--';
-  document.getElementById('mode-card-status').innerHTML = '';
-  document.getElementById('tank-temp-val').textContent = '--';
-  document.getElementById('tank-temp-status').textContent = '';
-  document.getElementById('tank-temp-message').textContent = '';
-  document.getElementById('tank-stat-energy').textContent = '--';
-  document.getElementById('tank-stat-greenhouse').textContent = '--';
-  var ghTrendResetEl = document.getElementById('tank-stat-greenhouse-trend');
-  if (ghTrendResetEl) ghTrendResetEl.innerHTML = '';
-  document.getElementById('inactive-modes').innerHTML = '';
-  resetLiveYesterdayHigh();
-  document.getElementById('graph-peak-label').textContent = "Yesterday's High: --";
-  var arc = document.getElementById('tank-gauge-arc');
-  if (arc) arc.setAttribute('stroke-dashoffset', '628');
-  // Clear component statuses
-  var compEls = document.querySelectorAll('.comp-status');
-  compEls.forEach(function(el) { el.textContent = '--'; });
-  // Clear simulation graph data and redraw empty canvas
-  timeSeriesStore.reset();
-  drawHistoryGraph();
-  // Clear the transition log — fetchLiveEvents() will repopulate it from the DB
-  transitionLog.length = 0;
-  resetEventsState();
-  var logsEl = document.getElementById('logs-list');
-  if (logsEl) {
-    logsEl.innerHTML = '<div data-empty="true" style="color:var(--on-surface-variant);font-size:13px;">Loading transitions…</div>';
-  }
-}
-
-function switchToSimulation() {
-  store.set('phase', 'simulation');
-  updateModeToggleUI(false);
-  persistModeInUrl('sim');
-  if (liveSource) liveSource.stop();
-  updateConnectionUI('disconnected');
-  clearLiveHistoryData();
-  balanceHistory = null;
-  balanceLivePoints = [];
-  balanceLiveEvents = [];
-  balanceLiveLastMode = null;
-  resetLiveYesterdayHigh();
-  renderBalanceCard();
-  stopStalenessTimer();
-  const banner = document.getElementById('staleness-banner');
-  if (banner) banner.classList.remove('visible');
-  updateConnectionOverlays();
-  updateDevicePushState();
-  updateSidebarSubtitle();
-  updatePhaseAwareCopy();
-}
-
-// Swap user-facing copy that differs between live and simulation modes,
-// so the Status view doesn't tell live users to "Adjust parameters in
-// Controls" or "Start the simulation".
-function updatePhaseAwareCopy() {
-  var isLive = store.get('phase') === 'live';
-  var desc = document.getElementById('status-view-description');
-  if (desc) {
-    desc.textContent = isLive
-      ? 'Live readings from the Shelly controller. Mode, valves, actuators, and 24h history.'
-      : 'Thermal simulation of the solar heating system. Adjust parameters in Controls and observe behavior here.';
-  }
-  var logsEmpty = document.getElementById('logs-list');
-  if (logsEmpty && logsEmpty.children.length <= 1) {
-    // Only replace the placeholder, not real log entries
-    var placeholder = logsEmpty.querySelector('div[data-empty]') || (logsEmpty.children.length === 0 ? null : logsEmpty.firstElementChild);
-    if (placeholder && placeholder.dataset && placeholder.dataset.empty === 'true') {
-      placeholder.textContent = isLive
-        ? 'No transitions yet. Awaiting controller activity…'
-        : 'No transitions yet. Start the simulation.';
-    }
-  }
-  var ctrl = document.getElementById('comp-controller');
-  if (ctrl && isLive) {
-    // In live mode the simulation "running" flag is meaningless. Show
-    // ACTIVE when the controller is responsive (will be re-evaluated by
-    // the next state broadcast).
-    ctrl.textContent = 'ACTIVE';
-  }
-}
-
-function updateSidebarSubtitle() {
-  var el = document.getElementById('sidebar-subtitle');
-  if (!el) return;
-
-  if (store.get('phase') === 'simulation') {
-    el.textContent = running ? 'Simulating...' : 'Ready';
-    el.className = '';
-    return;
-  }
-
-  // Live mode: reflect connection state
-  var displayState = getConnectionDisplayState();
-  switch (displayState) {
-    case 'active':
-      el.textContent = 'Live';
-      el.className = 'subtitle-live';
-      break;
-    case 'connecting':
-      el.textContent = 'Connecting\u2026';
-      el.className = 'subtitle-connecting';
-      break;
-    case 'device_offline':
-      el.textContent = 'Controller Offline';
-      el.className = 'subtitle-offline';
-      break;
-    case 'stale':
-      el.textContent = 'Stale';
-      el.className = 'subtitle-offline';
-      break;
-    case 'disconnected':
-    case 'never_connected':
-      el.textContent = 'Offline';
-      el.className = 'subtitle-offline';
-      break;
-  }
-}
 
 // ── State ──
 let config = null;
@@ -652,10 +259,11 @@ async function init() {
   updateDisplay(model.getState(), { mode: 'idle', valves: { vi_btm: false, vi_top: false, vi_coll: false, vo_coll: false, vo_rad: false, vo_tank: false, v_air: false }, actuators: { pump: false, fan: false, space_heater: false }, transition: null });
 
   // Initialize live/simulation mode toggle
+  initConnection({ setRunning: (v) => { running = v; } });
   initModeToggle();
   initDeviceConfig();
-  initRelayBoard({ getLiveSource: () => liveSource });
-  initWatchdogUI({ getLiveSource: () => liveSource });
+  initRelayBoard({ getLiveSource });
+  initWatchdogUI({ getLiveSource });
 
   // On deploys without live mode (e.g. GitHub Pages) load the
   // pre-baked simulation snapshot so the dashboard is populated on
